# auth-service/src/main/java/com/ecommerce/auth/dto/LoginRequest.java
package com.ecommerce.auth.dto;
import jakarta.validation.constraints.NotBlank;
import lombok.Data;
@Data public class LoginRequest { @NotBlank private String username; @NotBlank private String password; }